[feat] Added automatic login and guest users
Summary: All users of the site must now be logged in. If they have
not logged in via 3rd party authentication we log them in as guests.

All additions made in application_controller. If user has not logged
in, we log him in as a guest user. If a user logs manually we have the
option to import information from their temporary guest user to their
actual user record. When the user logs in the guest user record is
deleted. Views not have access to 2 new objects, @guest_user and
@current_or_guest_user. We have left @current_user untouched to allow
the conditional functionality within views to still work